refactor: Split the doctypes files
