+++ src/doctypes/index.js
@@ -1,0 +1,44 @@
+import {
+  ACCOUNT_DOCTYPE,
+  GROUP_DOCTYPE,
+  TRANSACTION_DOCTYPE,
+  SETTINGS_DOCTYPE,
+  BILLS_DOCTYPE
+} from './schema'
+
+export const offlineDoctypes = [
+  ACCOUNT_DOCTYPE,
+  GROUP_DOCTYPE,
+  TRANSACTION_DOCTYPE,
+  SETTINGS_DOCTYPE,
+  BILLS_DOCTYPE
+]
+
+export {
+  RECIPIENT_DOCTYPE,
+  ACCOUNT_DOCTYPE,
+  GROUP_DOCTYPE,
+  TRANSACTION_DOCTYPE,
+  SETTINGS_DOCTYPE,
+  BILLS_DOCTYPE,
+  TRIGGER_DOCTYPE,
+  APP_DOCTYPE,
+  KONNECTOR_DOCTYPE,
+  COZY_ACCOUNT_DOCTYPE,
+  PERMISSION_DOCTYPE,
+  BANK_ACCOUNT_STATS_DOCTYPE,
+  schema
+} from './schema'
+
+export { HasManyReimbursements, HasManyBills } from './relationships'
+
+export {
+  accountsConn,
+  groupsConn,
+  triggersConn,
+  transactionsConn,
+  appsConn,
+  billsConn,
+  settingsConn,
+  recipientsConn
+} from './querySpecs'
